Fix AWS on-prem snags & add docs (#517)
Addresses some snags when attempting to run the AWS binary on-prem, with local S3 and MySQL services, and provides a README to guide the user when attempting this.

The AWS issuer storage has been updated to allow for custom AWS SDK Config and S3 Options to be passed in, this is necessary to be able to stop it from moving the bucket name from the path of S3 requests to a new host prepended to the configured endpoint 🙄. This brings it into line with the Tessera AWS storage implementation which already does this.