Migrate to cloud-init (#591)
Migrates away from the to-be-deprecated GCE-just-run-a-VM option to use cloud-init instead.

The COS VM is configured using cloud-init such that:

a new systemd service which uses Docker to run the tesseract binary is created
the Docker container is named using the origin prefix for ease of grouping logging in GCP (using jsonPayload.container.name=... as a filter)

And the MIG terraform resource is now configured to cause the terraform command to wait for the managed instances to be successfully updated before returning, this means that it's no longer possible for a bad update to silently roll back.