Disable linter for x509 fork (#643)
No need to lint this when we don't want to make any changes to the fork.

Note that in the run it says `Flag --skip-dirs has been deprecated, check the documentation for more information.`

However, this works for now. The migration away from this is to set up a whole config file for the linter, which isn't on my short term roadmap.